refactor(sdk): rename auth.{py,ts} → credential.{py,ts} for naming consistency
The `Credential` type was already uniform, but the module holding it
diverged: Rust `rest::credential`, but Python/Node glue used `auth`.

Standardize on `credential` so the noun matches at every layer,
following the Azure/GCP two-level vocabulary:

- Module holding the Credential type = `credential` — mirrors
  `azure.core.credentials` / `google.auth.credentials`.
  Python `boxlite.auth` → `boxlite.credential`;
  Node `lib/auth.ts` → `lib/credential.ts`; Rust `rest::credential`
  already correct.
- CLI sign-in verb stays `boxlite auth login` (GCP `gcloud auth login`).
- On-disk store stays `credentials.toml` (AWS `~/.aws/credentials`).
- Type names `Credential`/`ApiKeyCredential`/`AccessToken` unchanged
  (Azure `TokenCredential` singular convention).

Pure path rename + import/doc sites. cargo check -p boxlite-python
-p boxlite-node clean. Python 123 unit + Node 73 unit green.
