+++ sdks/node/lib/credential.ts
