+++ sdks/node/lib/index.ts
@@ -35,7 +35,11 @@ export type { CopyOptions } from "./copy.js";
 
 // Credential abstraction (Azure-style: structural `Credential` interface
 // + concrete `ApiKeyCredential` class).
-export { ApiKeyCredential, type Credential, type AccessToken } from "./auth.js";
+export {
+  ApiKeyCredential,
+  type Credential,
+  type AccessToken,
+} from "./credential.js";
 
 // Export native module loader for advanced use cases
 export { getNativeModule, getJsBoxlite };
